new1

+++ Experiment5/Models/Employee.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public double Salary { get; set; }
+        public int Salary { get; set; }
         public bool Permanent { get; set; }
         public Department Department { get; set; }
         public List<Skill> Skills { get; set; }
